+++ mapper_000.h
@@ -10,7 +10,7 @@ class Cartridge;
 class Mapper_000 : public Mapper {
 
 public:
-  bool cpu_read_mapper(uint16_t adr, uint32_t &mapped_adr) override;
+  bool cpu_read_mapper(uint16_t adr, uint16_t &mapped_adr) override;
   bool cpu_write_mapper(uint16_t adr, uint32_t &mapped_adr) override;
   bool ppu_read_mapper(uint16_t adr, uint32_t &mapped_adr) override;
   bool ppu_write_mapper(uint16_t adr, uint32_t &mapped_adr) override;
